+++ pypesto/__init__.py
@@ -42,5 +42,6 @@ from . import logging
 from . import startpoint
 from . import store
 from . import visualize
+from . import C
 
 logging.log()
