+++ pypesto/ensemble/ensemble.py
@@ -2,7 +2,7 @@ import logging
 from functools import partial
 import numpy as np
 import pandas as pd
-from typing import Sequence, Tuple, Callable, Dict, List, Optional
+from typing import Sequence, Tuple, Union, Callable, Dict, List, Optional
 
 from .. import Result
 from ..objective import AmiciObjective
@@ -12,20 +12,20 @@ from ..engine import (
     MultiThreadEngine,
     SingleCoreEngine,
 )
-from ..predict import (
+from ..result import (
     PredictionConditionResult,
     PredictionResult,
 )
 from ..sample import geweke_test
+from ..C import (PREDICTOR, PREDICTION_ID, PREDICTION_RESULTS,
+                 PREDICTION_ARRAYS, PREDICTION_SUMMARY, OUTPUT,
+                 OUTPUT_SENSI, TIMEPOINTS, X_VECTOR, NX, X_NAMES,
+                 NVECTORS, VECTOR_TAGS, PREDICTIONS, MODE_FUN,
+                 EnsembleType, ENSEMBLE_TYPE, MEAN, MEDIAN,
+                 STANDARD_DEVIATION, SUMMARY, LOWER_BOUND,
+                 UPPER_BOUND, HISTORY, PERCENTILE,
+                 WEIGHTED_SIGMA)
 from .task import EnsembleTask
-from .constants import (PREDICTOR, PREDICTION_ID, PREDICTION_RESULTS,
-                        PREDICTION_ARRAYS, PREDICTION_SUMMARY, OUTPUT,
-                        OUTPUT_SENSI, TIMEPOINTS, X_VECTOR, NX, X_NAMES,
-                        NVECTORS, VECTOR_TAGS, PREDICTIONS, MODE_FUN,
-                        EnsembleType, ENSEMBLE_TYPE, MEAN, MEDIAN,
-                        STANDARD_DEVIATION, SUMMARY, LOWER_BOUND,
-                        UPPER_BOUND, get_percentile_label, HISTORY,
-                        WEIGHTED_SIGMA)
 
 logger = logging.getLogger(__name__)
 
@@ -1022,3 +1022,37 @@ def get_vector_indices(trace_start: np.ndarray,
         return candidates[indices.astype(int)]
     else:
         return candidates[:n_vectors]
+
+
+def get_percentile_label(percentile: Union[float, int, str]) -> str:
+    """Convert a percentile to a label.
+
+    Labels for percentiles are used at different locations (e.g. ensemble
+    prediction code, and visualization code). This method ensures that the same
+    percentile is labeled identically everywhere.
+
+    The percentile is rounded to two decimal places in the label representation
+    if it is specified to more decimal places. This is for readability in
+    plotting routines, and to avoid float to string conversion issues related
+    to float precision.
+
+    Parameters
+    ----------
+    percentile:
+        The percentile value that will be used to generate a label.
+
+    Returns
+    -------
+    The label of the (possibly rounded) percentile.
+    """
+    if isinstance(percentile, str):
+        percentile = float(percentile)
+        if percentile == round(percentile):
+            percentile = round(percentile)
+    if isinstance(percentile, float):
+        percentile_str = f'{percentile:.2f}'
+        # Add `...` to the label if the percentile value changed after rounding
+        if float(percentile_str) != percentile:
+            percentile_str += '...'
+        percentile = percentile_str
+    return f'{PERCENTILE} {percentile}'
