+++ pypesto/objective/priors.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 
 from .function import ObjectiveBase
 from .aggregated import AggregatedObjective
-from .constants import MODE_FUN, MODE_RES, FVAL, GRAD, HESS, RES, SRES, CHI2
+from ..C import MODE_FUN, MODE_RES, FVAL, GRAD, HESS, RES, SRES, CHI2
 from .util import res_to_chi2
 
 from .base import ResultDict
