+++ pypesto/petab/importer.py
@@ -14,7 +14,7 @@ from ..objective import AmiciObjective, AmiciObjectBuilder, AggregatedObjective
 from ..predict import AmiciPredictor, PredictionResult
 from ..predict.constants import CONDITION_SEP
 from ..objective.priors import NegLogParameterPriors, get_parameter_prior_dict
-from ..objective.constants import MODE_FUN, MODE_RES
+from ..C import MODE_FUN, MODE_RES
 from ..startpoint import FunctionStartpoints, StartpointMethod
 
 try:
