+++ pypesto/predict/__init__.py
@@ -8,9 +8,4 @@ optional post-processing.
 
 
 from .amici_predictor import AmiciPredictor
-from .result import (
-    PredictionResult,
-    PredictionConditionResult,
-)
 from .task import PredictorTask
-from . import constants
