+++ test/petab/test_sbml_conversion.py
@@ -36,7 +36,7 @@ class AmiciObjectiveTest(unittest.TestCase):
 
             df = objective.check_grad(
                 x0, eps=1e-3, verbosity=0,
-                mode=pypesto.objective.constants.MODE_FUN)
+                mode=pypesto.C.MODE_FUN)
             print("relative errors MODE_FUN: ", df.rel_err.values)
             print("absolute errors MODE_FUN: ", df.abs_err.values)
             assert np.all((df.rel_err.values < RTOL) |
@@ -44,7 +44,7 @@ class AmiciObjectiveTest(unittest.TestCase):
 
             df = objective.check_grad(
                 x0, eps=1e-3, verbosity=0,
-                mode=pypesto.objective.constants.MODE_RES)
+                mode=pypesto.C.MODE_RES)
             print("relative errors MODE_RES: ", df.rel_err.values)
             print("absolute errors MODE_RES: ", df.rel_err.values)
             assert np.all((df.rel_err.values < RTOL) |
